First iteration of refactor
All commands sorted into cogs. Leaderboards and Members thereof are now implemented as classes, and the rate limiting of API calls is now handled on a per-guild basis, so no issues may be caused by a recent call in one guild causing a call in another to fail unexpectedly. I am 100% happy with both info_commands and config_commands. I am still debating whether or not make_embed should be a method of the Leaderboard class. On the one hand, it would make the structure of the code much more obvious. On the other hand, I would have to create two essentially identical methods to handle making an embed from the cached leaderboard data and from newly requested data.